+++ src/FhirStore.CommonVersioned/Storage/IVersionedResourceStore.cs
@@ -72,10 +72,12 @@ public interface IVersionedResourceStore : IResourceStore, IDisposable, IReadOnl
         string id,
         HashSet<string> protectedResources);
 
-    /// <summary>Performs a type search in this resource store.</summary>
-    /// <param name="parameters">The query.</param>
+    /// <summary>
+    /// Performs a type search in this resource store.
+    /// </summary>
+    /// <param name="parameters">The search parameters.</param>
     /// <returns>
-    /// An enumerator that allows foreach to be used to process type search in this collection.
+    /// An enumerator that allows foreach to be used to process the search results in this collection.
     /// </returns>
     IEnumerable<Hl7.Fhir.Model.Resource>? TypeSearch(IEnumerable<ParsedSearchParameter> parameters);